Review + apply best practices
- split interface and impl
- fix readme
- align name/nameHint in Config (so that it shows up nicely in Apache Felix WebConsole)
- simplify meter Component Name (otherwise it would cause too many subfolders in Apache Felix Configuration Admin directory)
