+++ io.openems.edge.meter.siemens/bnd.bnd
@@ -9,6 +9,7 @@ Bundle-Version: 1.0.0.${tstamp}
 	io.openems.common,\
 	io.openems.edge.bridge.modbus,\
 	io.openems.edge.common,\
-	io.openems.edge.meter.api
+	io.openems.edge.meter.api,\
+
 -testpath: \
 	${testpath}
