+++ io.openems.edge.meter.siemens/readme.adoc
@@ -1,7 +1,8 @@
-= Janitza Meters UMG 96RM-E | UMG 604
+= Siemens Meters PAC2200 / 3200 / 4200
 
-Implemented Natures::
+Implemented Natures
+
 - SymmetricMeter
 - AsymmetricMeter
 
-https://github.com/OpenEMS/openems/tree/develop/io.openems.edge.meter.janitza[Source Code icon:github[]]
+https://github.com/OpenEMS/openems/tree/develop/io.openems.edge.meter.siemens[Source Code icon:github[]]
