+++ io.openems.edge.meter.siemens/src/io/openems/edge/meter/siemens/SiemensPacMeter.java
@@ -1,0 +1,25 @@
+package io.openems.edge.meter.siemens;
+
+import io.openems.edge.bridge.modbus.api.ModbusComponent;
+import io.openems.edge.common.channel.Doc;
+import io.openems.edge.common.component.OpenemsComponent;
+import io.openems.edge.common.modbusslave.ModbusSlave;
+import io.openems.edge.meter.api.AsymmetricMeter;
+import io.openems.edge.meter.api.SymmetricMeter;
+
+public interface SiemensPacMeter
+		extends SymmetricMeter, AsymmetricMeter, ModbusComponent, OpenemsComponent, ModbusSlave {
+
+	public enum ChannelId implements io.openems.edge.common.channel.ChannelId {
+		;
+		private final Doc doc;
+
+		private ChannelId(Doc doc) {
+			this.doc = doc;
+		}
+
+		public Doc doc() {
+			return this.doc;
+		}
+	}
+}
